+++ app/models/folio/application_record.rb
@@ -4,6 +4,7 @@ class Folio::ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
 
   include Folio::Filterable
+  include Folio::HasSanitizedFields
   include Folio::NillifyBlanks
   include Folio::RecursiveSubclasses
   include Folio::Sortable