+++ app/models/folio/user.rb
@@ -5,6 +5,8 @@ class Folio::User < Folio::ApplicationRecord
   include Folio::HasAddresses
   include Folio::HasNewsletterSubscription
 
+  has_sanitized_fields :email, :first_name, :last_name, :nickname
+
   # used to validate before inviting from console in /console/users/new
   attribute :skip_password_validation, :boolean, default: false